transaction/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
paramiko
pymysql
flask_cors